Add move the controller test to the right folder
